Bug 2071196 - Avoid 32-bit overflow when converting chunk indices to file offsets in CacheFile (Coverity CID 1286534) r=necko-reviewers,valentin
kChunkSize is an int32_t and the chunk index is a uint32_t, so
`aIndex * kChunkSize` is evaluated in 32 bits and wraps once the index reaches
16384, i.e. for entries larger than 4GB. Both results feed int64_t offsets.
CacheFile::BytesFromChunk already casts before multiplying; do the same at the
two remaining sites.

Differential Revision: https://phabricator.services.mozilla.com/D325296

+++ netwerk/cache2/CacheFile.cpp
@@ -1624,7 +1624,8 @@ nsresult CacheFile::GetChunkLocked(uint32_t aIndex, ECallerType aCaller,
         if (startChunk != aIndex) {
           // Make sure the file contains zeroes at the end of the file
           rv = CacheFileIOManager::TruncateSeekSetEOF(
-              mHandle, startChunk * kChunkSize, aIndex * kChunkSize, nullptr);
+              mHandle, int64_t(startChunk) * kChunkSize,
+              int64_t(aIndex) * kChunkSize, nullptr);
           NS_ENSURE_SUCCESS(rv, rv);
         }
 
@@ -1952,7 +1953,7 @@ int64_t CacheFile::BytesFromChunk(uint32_t aIndex, bool aAlternativeData) {
   // theoretic bytes in advance
   int64_t advance = int64_t(i - aIndex) * kChunkSize;
   // real bytes till the end of the file
-  int64_t tail = dataSize - (aIndex * kChunkSize);
+  int64_t tail = dataSize - (int64_t(aIndex) * kChunkSize);
 
   return std::min(advance, tail);
 }
